Site Editor: Publish Patterns page to non-block themes
